Ignore missing chrome.manifest

+++ Makefile
@@ -4,14 +4,14 @@ all: xpi
 
 xpi: buildscript/makexpi.sh
 	cp buildscript/makexpi.sh ./
-	rm chrome.manifest
+	if [ -f chrome.manifest ]; then rm chrome.manifest; fi
 	cp jar-chrome.manifest chrome.manifest
 	./makexpi.sh -n $(PACKAGE_NAME)
 	rm ./makexpi.sh
 
 omnixpi: buildscript/makexpi.sh
 	cp buildscript/makexpi.sh ./
-	rm chrome.manifest
+	if [ -f chrome.manifest ]; then rm chrome.manifest; fi
 	cp omnixpi-chrome.manifest chrome.manifest
 	./makexpi.sh -n $(PACKAGE_NAME)
 	rm ./makexpi.sh -o
